requirements.txt
awscli==1.14.2
botocore==1.8.6
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.7
Cython==0.27.3
docutils==0.14
fasttext==0.8.3
future==0.16.0
idna==2.6
jmespath==0.9.3
nltk==3.2.5
numpy==1.14.0
pyasn1==0.4.2
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.12
six==1.11.0
textblob==0.15.0
tzlocal==1.5.1
urllib3==1.22
